+++ molsym/salcs/__init__.py
@@ -2,8 +2,10 @@ from . import projection_op
 from . import cartesian_coordinates
 from . import internal_coordinates
 from . import spherical_harmonics
+from . import salc_tools
 
 from .projection_op import ProjectionOp
+from .projection_op import ProjectOnObject
 from .cartesian_coordinates import CartesianCoordinates
 from .internal_coordinates import InternalCoordinates
 from .spherical_harmonics import SphericalHarmonics
