+++ molsym/salcs/projection_op.py
@@ -66,7 +66,7 @@ def eckart_conditions(symtext, translational=True, rotational=True):
     else:
         raise Exception("Calling this function is rather silly if you don't want either output...")
 
-def ProjectionOp(symtext, fxn_set, project_Eckart=True, manual_proj = None):
+def ProjectionOp(symtext, fxn_set, project_Eckart=True):
     """
     Projection operator: projects the functions in fxn_set into SALCs.
 
@@ -78,68 +78,89 @@ def ProjectionOp(symtext, fxn_set, project_Eckart=True, manual_proj = None):
     salcs = SALCs(symtext, fxn_set)
     orthogonalize = False
     #orthogonalize = True
-    if type(manual_proj) is not None:    
+    if manual_proj is None:    
+    for ir, irrep in enumerate(symtext.irreps):
+        if symtext.pg.is_linear:
+            irrmat = None
+        else:
+            irrmat = symtext.irrep_mats[irrep.symbol]
+        for se_fxn_set in fxn_set.SE_fxns:
+            equivcoord = min(se_fxn_set)
+            salc = np.zeros((irrep.d, irrep.d, numred))
+            if symtext.complex:
+                salc = np.zeros((irrep.d, irrep.d, numred), dtype=np.complex128)
+            for sidx in range(len(symtext)):
+                salc = fxn_set.special_function(salc, equivcoord, sidx, irrmat)
+            salc *= irrep.d/symtext.order
+            
+            # Project out Eckart conditions when constructing SALCs of Cartesian displacements
+            if isinstance(fxn_set, CartesianCoordinates) and project_Eckart:
+                orthogonalize = True
+                eckart_cond = eckart_conditions(symtext)
+                for i in range(irrep.d):
+                    for j in range(irrep.d):
+                        if not np.allclose(salc[i,j,:], np.zeros(salc[i,j,:].shape), atol=salcs.tol):
+                            salc[i,j,:] = project_out_Eckart(eckart_cond, salc[i,j,:])
+            # Convert complex SALCs to real
+            if symtext.complex and irrep.d==2:
+                nf =  1/np.sqrt(2)
+                new_i = nf * salc[:,0,:]+salc[:,1,:]
+                new_j = nf * (salc[:,0,:]-salc[:,1,:])/1j
+                salc[:,0,:] = new_i
+                salc[:,1,:] = new_j
+            # Add SALCs to SALC object
+            for i in range(irrep.d):
+                for j in range(irrep.d):
+                    if not np.allclose(salc[i,j,:], np.zeros(salc[i,j,:].shape), atol=salcs.tol):
+                        gamma = 1.0/np.linalg.norm(salc[i,j,:])
+                        salc[i,j,:] = molsym.symtools.normalize(salc[i,j,:])
+                        s = SALC(salc[i,j,:], irrep, equivcoord, i, j, gamma)
+                        salcs.addnewSALC(s, ir)
+    if symtext.complex:
+        remove_complexity = True
+    else:
+        remove_complexity = False
+    # Build convenience SALC data structures
+    salcs.finish_building(orthogonalize=orthogonalize, remove_complexity=remove_complexity)
+    return salcs
+
+def ProjectOnObject(symtext, fxn_set, project_Eckart=True, manual_proj):
+    """
+    Projection operator: projects the functions in fxn_set into SALCs.
+
+    :type symtext: molsym.Symtext
+    :type fxn_set: molsym.FunctionSet
+    :rtype: molsym.SALCs
+    """
+    numred = len(fxn_set)
+    salcs = SALCs(symtext, fxn_set)
+    orthogonalize = False
+    #orthogonalize = True
+    list_salc_ids = []
+    index_for_irrep = []
+    print("Projecting on user-specified SALCs to determine what irreps they belong to")
+    for r, row in enumerate(manual_proj):
         for ir, irrep in enumerate(symtext.irreps):
             if symtext.pg.is_linear:
                 irrmat = None
             else:
                 irrmat = symtext.irrep_mats[irrep.symbol]
             for se_fxn_set in fxn_set.SE_fxns:
-                equivcoord = min(se_fxn_set)
-                salc = np.zeros((irrep.d, irrep.d, numred))
-                if symtext.complex:
-                    salc = np.zeros((irrep.d, irrep.d, numred), dtype=np.complex128)
-                for sidx in range(len(symtext)):
-                    salc = fxn_set.special_function(salc, equivcoord, sidx, irrmat)
-                salc *= irrep.d/symtext.order
-                
-                # Project out Eckart conditions when constructing SALCs of Cartesian displacements
-                if isinstance(fxn_set, CartesianCoordinates) and project_Eckart:
-                    orthogonalize = True
-                    eckart_cond = eckart_conditions(symtext)
-                    for i in range(irrep.d):
-                        for j in range(irrep.d):
-                            if not np.allclose(salc[i,j,:], np.zeros(salc[i,j,:].shape), atol=salcs.tol):
-                                salc[i,j,:] = project_out_Eckart(eckart_cond, salc[i,j,:])
-                # Convert complex SALCs to real
-                if symtext.complex and irrep.d==2:
-                    nf =  1/np.sqrt(2)
-                    new_i = nf * salc[:,0,:]+salc[:,1,:]
-                    new_j = nf * (salc[:,0,:]-salc[:,1,:])/1j
-                    salc[:,0,:] = new_i
-                    salc[:,1,:] = new_j
-                # Add SALCs to SALC object
-                for i in range(irrep.d):
-                    for j in range(irrep.d):
-                        if not np.allclose(salc[i,j,:], np.zeros(salc[i,j,:].shape), atol=salcs.tol):
-                            gamma = 1.0/np.linalg.norm(salc[i,j,:])
-                            salc[i,j,:] = molsym.symtools.normalize(salc[i,j,:])
-                            s = SALC(salc[i,j,:], irrep, equivcoord, i, j, gamma)
-                            salcs.addnewSALC(s, ir)
-    else:
-        print("Projecting on user-specified SALCs to determine what irreps they belong to")
-        for row in manual_proj:
-            for ir, irrep in enumerate(symtext.irreps):
-                if symtext.pg.is_linear:
-                    irrmat = None
+                if np.any(np.abs(row[se_fxn_set])) > 1e-4: 
+                    big_salc = np.zeros((irrep.d, irrep.d, numred))
+                    for equivcoord in se_fxn_set:
+                        salc = np.zeros((irrep.d, irrep.d, numred))
+                        if symtext.complex:
+                            salc = np.zeros((irrep.d, irrep.d, numred), dtype=np.complex128)
+                        for sidx in range(len(symtext)):
+                            salc = fxn_set.special_function(salc, equivcoord, sidx, irrmat)
+                        salc *= irrep.d/symtext.order
+                        big_salc += salc * row[equivcoord]
+                    if (np.abs(big_salc) > 1e-6).any():
+                        list_salc_ids.append(irrep.symbol)
+                        index_for_irrep.append(r)
                 else:
-                    irrmat = symtext.irrep_mats[irrep.symbol]
-                for se_fxn_set in fxn_set.SE_fxns:
-                    if np.any(np.abs(row[se_fxn_set])) > 1e-4: 
-                        big_salc = np.zeros((irrep.d, irrep.d, numred))
-                        for equivcoord in se_fxn_set:
-                            salc = np.zeros((irrep.d, irrep.d, numred))
-                            if symtext.complex:
-                                salc = np.zeros((irrep.d, irrep.d, numred), dtype=np.complex128)
-                            for sidx in range(len(symtext)):
-                                salc = fxn_set.special_function(salc, equivcoord, sidx, irrmat)
-                            salc *= irrep.d/symtext.order
-                            big_salc += salc * row[equivcoord]
-                        if (np.abs(big_salc) > 1e-6).any():
-                            list_salc_ids.append(irrep.symbol)
-                    else:
-                        continue
-
+                    continue
     if symtext.complex:
         remove_complexity = True
     else:
